I'm afraid the new design wouldn't perform very well. E.g. if we implement the number formatting features of printf() on top if it, it would probably be necessary to clone a BNumberFormatImpl for each field and since the latter is an object of a class provided by an add-on, it can be arbitrarily complex.
So here's a newer design: The B*FormatImpl classes are re-entrant
and its instances immutable, which get their parameters as objects via the
Format()/Parse() methods. The B*Format classes are still convenient
user API frontends with setters for the formatting/parsing parameters,
but no longer as light-weight as before, since they do contain the
parameters. When formatting with non-default parameters a
B*Format needs to be cloned, but this can be done on the stack and should
be reasonably fast (at least the overhead is well known).

Moreover BNumberFormat is no longer responsible for formatting all kinds
of numbers. Instead the subclasses BFloatFormat and BIntegerFormat are
used for the respective kinds of numbers.

B*FormatParameter classes encapsulate the parameters for the corresponding
formats.